+++ bookkeeping/src/sdk/models/operations/get_benefits_enrollment_id_.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import enrollment as shared_enrollment
+from ...models.shared import enrollment as shared_enrollment
 from typing import Optional
 
 
@@ -19,11 +19,11 @@ class GetBenefitsEnrollmentIDRequest:
 class GetBenefitsEnrollmentIDResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     enrollment: Optional[shared_enrollment.Enrollment] = dataclasses.field(default=None)
     r"""An Enrollment records"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""