+++ bookkeeping/RELEASES.md
@@ -9,3 +9,64 @@ Based on:
 - [python v1.0.0] bookkeeping
 ### Releases
 - [PyPI v1.0.0] https://pypi.org/project/wingspan_bookkeeping/1.0.0 - bookkeeping
+
+
+## 2023-11-13 01:29:56
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.0] bookkeeping
+### Releases
+- [PyPI v2.0.0] https://pypi.org/project/wingspan_bookkeeping/2.0.0 - bookkeeping
+
+## 2023-11-20 01:30:38
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.1] bookkeeping
+### Releases
+- [PyPI v2.0.1] https://pypi.org/project/wingspan_bookkeeping/2.0.1 - bookkeeping
+
+## 2023-11-27 01:30:23
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.2] bookkeeping
+### Releases
+- [PyPI v2.0.2] https://pypi.org/project/wingspan_bookkeeping/2.0.2 - bookkeeping
+
+## 2023-12-11 01:30:59
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.3] bookkeeping
+### Releases
+- [PyPI v2.0.3] https://pypi.org/project/wingspan_bookkeeping/2.0.3 - bookkeeping
+
+## 2023-12-18 01:30:55
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.126.4 (2.214.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.1.0] bookkeeping
+### Releases
+- [PyPI v2.1.0] https://pypi.org/project/wingspan_bookkeeping/2.1.0 - bookkeeping
+
+## 2024-01-08 01:31:07
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.133.1 (2.228.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.0] bookkeeping
+### Releases
+- [PyPI v3.0.0] https://pypi.org/project/wingspan_bookkeeping/3.0.0 - bookkeeping